fixes issue 424
short-circuit event delivery for the following events
by not calling subsequent callbacks on the first callback
that asks to not deliver to the app:

+ restore_state_ex
+ pre_syscall
+ exception
+ signal

note that this does violate the "priority order" of events where the last
one is supposed to have final say b/c it won't even see the event: but only
one registrant should "own" an event in any case.

note also that all but signal events were already implemented as a
short-circuit, so this just makes it official and documents it.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@685 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6