Fix client.thread test to use right operand size.
movzb will only copy the low byte out of idx, when idx is a 32-bit integer.
This creates a problem if the offset returned by dr_raw_tls_calloc() is greater
than 255.  On X64 32-bit registers are automatically zero extended, so we don't
need to do anything special to get a zero-extend.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@774 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6