i#336: Adjust signal info arrays to include SIGRTMAX.
This patch creates a SIGARRAY_SIZE macro that is large enough to hold
all the valid signals.  MAX_SIGNUM is now considered to be equal to the
largest valid signal number, so it is one less than SIGARRAY_SIZE.  This
change canonicalizes all of the loops that iterate over signals to
iterate over the inclusive range [1..MAX_SIGNUM].  We should always
follow this pattern when iterating over signal numbers to prevent bugs.

This patch also fixes a bug where we were not noticing that we had signals
pending after unblocking them on a sigreturn.  This happens when a nested signal
is taken, but the application blocks it during the execution of the active
handler.  When we return, we need to see if any of the pending signals have been
unblocked, and if so, set dcontext->signals_pending to true so that it will be
delivered later from dispatch.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@992 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6