Implement drsyms for Linux.
Partially implements i#445.  We still have open issues.  In particular,
the most glaring is that symbol lookup uses strstr to find a substring.
This allows us to lookup mangled C++ symbols, but falls over if the
substring is not unique to a function name, as in malloc (i#581).

Fixing the substring matching really depends on name mangling (i#545) so
we can get a mangled symbol to look up, which I intend to tackle next,
since it's part of libelftc.

This also does a poor job of following debuglink, i#576.

We also intend to change the API to not require the modname! prefix in
the future, which is i#574.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1041 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6