Stop using strip -p to preserve mtimes.
Binutils strip doesn't preserve the mtime with enough resolution for our
purposes.  Because the link step takes less than a second for our
binaries and DSOs, make sees an mtime on the .so file that is earlier
than the one on the input .o files, and will re-link every binary and
DSO on a second incremental build.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1098 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6