Merge empty regions in /proc/pid/maps into the next region.
Fixes issue 599.

We've only ever seen these empty regions at the stack guard pages, such
as in i#366 and i#599.  Sometimes the stack region following the empty
region begins at the same start address, as in i#366, and sometimes it
starts a page afterwards, as in i#599.  We handle both cases now.

Tested by running this on the linux-x64-lucid builder.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1131 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6