Suffix failing test names with _FLAKY.
We're not interested in fixing these tests at the moment, and if we
disable them, they are unlikely to ever be fixed in the future.  Our
compromise is to keep running them, but mark them as flaky so we can
turn the bot green and easily detect any regressions in non-FLAKY tests.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1187 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6