Check the private PEB heap before calling NtQueryInformationProcess
While working on i#784 I attached to unit_tests.exe while the loader was
running a couple of times and this showed up once or twice.  dbghelp
makes lots of heap allocations and hits this a lot.  I didn't actually
measure, but it seems faster to try the pointer deref + compare before
doing the NtQueryInformationProcess syscall.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1283 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6