Fix building with DEBUG and without DEADLOCK_AVOIDANCE.
While debugging DrMemory i#784 I wanted a build with logging but without
anything potentially slow, so I turned off this as well as DEBUG_MEMORY
etc.  I had to make these changes in order to build with DEBUG but
without DEADLOCK_AVOIDANCE.

In particular, it looks like change_linking_lock went from a mutex to a
recursive lock, and OWN_MUTEX does the right thing for both ifdef
DEADLOCK_AVOIDANCE but not without it.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1285 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6