Make module_handle_t opaque again, this time fixing the Windows build.
module_handle_t is now a pointer to an incomplete struct type that we
never define, like so:
  struct _module_handle_t;
  typdef struct _module_handle_t *module_handle_t;

This turns the following into a compiler error:
  module_data_t *mod = dr_lookup_module(pc);
  void (*proc)() = dr_get_proc_address(mod/*should have ->handle*/, "proc_name");

Currently, for this code we attempt to use mod's address to look up a
module and then find its dynamic symbol table, and usually we fail and
return NULL.  Getting a compiler error is nicer.

This change does *not* break ABI compatibility, but is source
incompatible.  If the user is passing mod->start or some other pointer
to dr_get_proc_address(), they will get a compiler error when they build
with the new headers.  The API doesn't imply support for that, and I
don't think it should, so I'm OK with making that change.

Internally, we have a lot of code that passes HMODULE, HANDLE, app_pc,
byte*, and void* values to routines that take a module_handle_t, so we
moved them to a new module_base_t type which is just a void*.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1328 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6