Import standalone sscanf implementation from old Kerberos code.
This is needed on modern Ubuntu because sscanf calls malloc and is no
longer re-entrant.  (issue 762)

The only change to the original was to reindent the file using four
spaces instead of tabs.  Further changes are necessary to make this
useful to DR, and will be in a separate diff, to make the origin of the
functional changes clear.

Retreived from:
https://github.com/krb5/krb5/blob/4732e33646e665194630a9a5fe4d91c1c900ac14/src/lib/krb5/posix/sscanf.c