i#773: Fix racy assertion in thread takeover.
Fixes issue 773.

The race here was that we'd actually find our takeover record, and then
check num_takeover_records, which might have been updated after we
signalled the initiating thread.  This could have been fixed by making a
local variable copy, but I think grabbing the event, checking whether it
was found, and doing the signal outside the loop is cleaner.

There was also a boolean logic error where I used an || instead of &&.
This would have lead to bad behavior when receiving signals (suspend or
others) on a thread while it was still initializing.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1370 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6