i#531: Batch up add-symbol-file command prints and support gdb 7.2.
The main change is to print the add-symbol-file commands for the client
and its immediate dependencies in a single syslog that the user can copy
and paste into gdb.  -privload_register_gdb no longer hides the
add-symbol-file commands.

The Python script now checks for gdb 7.3 or more recent, and only tries
to set the breakpoint for gdb 7.3 or newer.

Automatically adding symbol files on attach in the Python script will be
implemented in a future patch.

Review URL: https://codereview.appspot.com/6216051

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1378 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6