i#239: Synch with all other threads on fork.
This helps to ensure that all DR locks are released when we begin
execution in the child.  If the app is spawning threads and forking
concurrently, then we may fail to synch with a recently spawned thread
due to issue 26, and so we may still have possible leaks and deadlocks
in the child process.

Review URL: https://codereview.appspot.com/6247048

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1382 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6