i#46: Linux libc isolation: memcpy, memset
This prevents the application from intercepting DR's calls to memcpy and
memset.

On x64, the performance is not much different from glibc.  However, on
x86_32, the perfomance is ~2x worse than glibc.  The gcc-optimized loop
in C performs much worse than the 'rep movs word' approach.

glibc uses SSE3 instructions on my workstation.  If we want more
performance, we can go down this road, but I'd prefer to avoid the
complexity, given that it's been a source of bugs for us in the past.