i#46: Avoid libc stdout, stderr, and stdin imports for _fileno.
There should be no behavior change for processes launched under DR.  The
only conforming way to change _fileno is with freopen, which closes the
original fd.  This triggers our client private fd behavior which dups
the fd.

If we implement attach, there may be a problem if the app has reopened
the stdio streams with different fds.