Merge all unit tests into one unit_tests binary.
This reduces the time to run the suite considerably, especially on Windows
where compilation is slower.
  merged suite time:    6m52.531s
  isolated suite time:  8m43.022s

Now it is possible to set -DSTANDALONE_UNIT_TEST=ON in a normal build dir and
build unit_tests along with DR.  It defaults to off, because it re-compiles all
of DR's source files.

Review URL: https://codereview.appspot.com/6349094

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1463 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6