i#46: Environment isolation.
This cuts the setenv, getenv, and __environ imports from libc.  Instead, we get
the envp pointer when the loader calls our _init() routine.  Depending on how
libc implements setenv, we may or may not see the app updating its environment
as the app runs.  All of our important reads happen early, so this is OK.