i#46: Avoid libc stdout, stderr, and stdin imports for _fileno.
There should be no behavior change for processes launched under DR.  The
only conforming way to change _fileno is with freopen, which closes the
original fd.  This triggers our client private fd behavior which dups
the fd.

If we implement attach, there may be a problem if the app has reopened
the stdio streams with different fds.

Review URL: https://codereview.appspot.com/6408046

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1467 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6