i#622: IR accessor inlining.
This is just a proof of concept for inlining some frequently called
small functions.  If this proves to work well with different versions of
gcc and cl.exe and Windows and Linux, we can use this as a basis for
adding more inline functions.

This gives a barely measurable 1% improvement for DR on DumpRenderTree
and for Dr. Memory on googleurl_unittests.