i#853: Use all_memory_areas in performance critical places.
This can be disabled if it causes problems with -no_use_all_memory_areas.  We can expand this option to cover more memory queries if they become performance bottlenecks.

This improves performance on DRT by 25%, and passes the suite on Ubuntu 12.04, Fedora 13, and Fedora 17.

Review URL: https://codereview.appspot.com/6449045

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1506 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6