i#622: IR accessor inlining.
This is just a proof of concept for inlining some frequently called
small functions.  If this proves to work well with different versions of
gcc and cl.exe and Windows and Linux, we can use this as a basis for
adding more inline functions.

To support widely used versions of gcc, routines that pass or return structs by value are implemented using macros for better performance.

This gives a 6% improvement for DR on DumpRenderTree.

Review URL: https://codereview.appspot.com/6419062

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1511 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6