Split privload_map_and_relocate() into map_elf_phdrs() and find_pt_interp()
map_elf_phdrs() is responsible for taking an ELF, parsing its program
headers, and mapping in the PT_LOAD segments.  There is no reason it
has to be concerned with reading e_entry and PT_INTERP.  We can get
those things later by parsing the mapped in image.  This avoids the
_IF_LINUX() parameters that the private loader doesn't need.

Unfortunately, for now, map_elf_phdrs() still needs a text_addr out
param because the section headers are generally not mapped in.

Review URL: https://codereview.appspot.com/6459045

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1513 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6