i#46: Avoid common libc imports in a release build.
Turning on optimizations causes tolower to be inlined, which creates a
__ctype_tolower_loc import.  For a variety of common string routines,
when the destination buffer size is known at compile time but the copy
size is not, gcc will call a __*_chk variant of the routine.  We just
alias these versions to the originals and rely on caller cleanup of
args.

Review URL: https://codereview.appspot.com/6446089

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1516 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6