i#47: Use an ELF entry point instead of LD_PRELOAD for early injection
By adding an entry point, we can invoke libdynamorio.so as an
executable.  It is still marked as type ET_DYN, but the kernel will go
ahead and load us in the usual way.  We get to look at argc, argv, envp,
and auxv first, and fix them up to match what the kernel would have done
for the app.

It is now possible to use early injection to run both statically and
dynamically linked apps.

Instead of moving strings around on the stack in DR, drloader now
adjusts the argv array before calling execve.  This costs an extra
execve() during startup, but it's a little cleaner than moving strings
around.  Eventually drloader and drrun should be combined, and we can
avoid the extra execve() initialization.

Review URL: https://codereview.appspot.com/6450102

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1550 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6