i#47: Manage LDT entries on ia32 during early injection
During early injection, the app doesn't have any segments set up.  We
request two segments, one for DR, and another for our private libraries
(LIB_SEG_TLS).  When the app (usually via ld.so) makes its initial call
to set_thread_area, we return the selector that we got for our private
libraries.

Previously we'd assert if the app didn't have TLS.

We also now do a little better at handling app segment updates (i#107).
Previously we'd mangle a segment update to a nop, but now we insert a
clean call to instruct DR to use the base from the appropriate LDT entry
that we've saved.