i#891: Add dr_get_main_module() to the API
This provides a reliable way for the client to get information about the
main app executable.  The existing pattern for getting the main module
is:

  dr_lookup_module_by_name(dr_get_application_name());

However, the above is unreliable if there are multiple modules with the
same preferred name.  More likely, on Linux, dr_get_application_name()
relies on the proc filesystem, which could be unavailable for various
reasons.  This API is more robust in the face of such corner cases.

Review URL: https://codereview.appspot.com/6492106

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1589 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6