i#893: add /safeseh:no
Added /safeseh:no for the core and for all the tests in order to
avoid errors from VS2012 due to our asm code.
Had to move core unit_tests down below the definition of LD_FLAGS.
Long-term we should make tools.asm SEH-compliant, so not resolving
the issue with this workaround.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1594 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6