i#928: Reset bb->exit_target when re-building sandboxed bbs
In issue 577 and r1047, we reset bb->exit_type, but we didn't reset
bb->exit_target.

Normally exit_target will get overwritten if we end on a cti, but if we
have more than 5 stores in the bb, we will stop early on a non-cti
instruction to avoid overly-long sandboxed bbs.  As a result, we attempt
to terminate the bb by jumping to exit_target instead of doing a
fall-through.

Fixes issue 928.

Review URL: https://codereview.appspot.com/6614050

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1644 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6