i#940: Always use a thread private copy for trace building
The underlying issue is that fragment recreation is not deterministic in
the face of transitions between sandboxing and readonly pages in the
code cache.

This change side steps the problem by building a temp thread private
fragment earlier, and using that to extend the trace instead of the
existing thread-shared fragment.

There should be no performance impact, because this just moves work that
we were already doing earlier.

Review URL: https://codereview.appspot.com/6634049

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1653 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6