fixes issue 946
When translating the state of a selfmod fragment, recreate from the copy of
the app code, pretending to be at the original pc (so low-2GB
instrumentation differences don't surface).

Added a fault test to the selfmod test.  As part of this added sharing of
signal handler code and sigsetjmp throughout the suite.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1656 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6