fixes issue 936
If we can't allocate reachable memory for landing pads, we use the padding
at the end of ntdll's +rx segment.  This gets mixed-mode working on win8
and is sufficient for now, but it's fragile: there might not be enough
padding.  Long-term we'll need to borrow space from multiple areas in
ntdll.

Adds trimming of landing pads to cut space needed in half.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1659 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6