fixes issue 949
Support 3rd-party extensions by requesting that the client use
find_package() to locate the extension or collection of extensions.
The resulting config file will set up the imported extension targets, which
must be libraries with names matching the extensions.  It also sets
DynamoRIO_EXT_${extname}_INC to point at the include directory.  This
allows the client to then use the same use_DynamoRIO_extension() function
as for DR-packaged extensions.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1671 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6