fixes issue 955
Adds the notion of a "<client_basename>.drpath" file: a text file that
contains a newline-separated list of search paths.  The paths are
absolute-only for now.  On Windows, if DynamoRIO_RPATH is ON,
configure_DynamoRIO_client() creates this file, and
use_DynamoRIO_extension() appends to it.  The private loader on both
plaforms reads the file.  This makes it easier to use third-party
Extensions.

Added documentation and a release note item for this feature.  Tested on
Dr. Memory's drsyscall test.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1674 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6