fixes issue 984
We cannot use whether a system library has been loaded as a condition
for whether to swap the state, as we need to make that decision at
gencode creation time.  Rather than going back and re-generating
gencode and handling complexities of changing strategies later, I
downgraded the optimization to just check whether there's a client
library.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1727 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6