i#989: Treat fallthroughs as transfers in check_new_page_contig()
This prevents building bbs that fall through from a readonly vmarea into
a sandboxed area, as well as other incompatible area transitions.  If
the first instr of the bb is straddling a vmarea, we let bb building
continue with the new vmarea flags so we don't create an empty bb.  This
case is covered by the selfmod_last_byte() test.

Fixes issue 989.

Review URL: https://codereview.appspot.com/6843126

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1751 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6