Don't add -O2 to clients in DynamoRIOConfig.cmake
Changed CLIENT_DEBUG to OPTIMIZE_SAMPLES, so we have a way of independently turning on optimizations for samples.

This fixes the drmemory build by silencing a warning that's only present
with -O2, but not -O3, which is what drmemory builds with in release
build.

Review URL: https://codereview.appspot.com/6902050

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1760 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6