i#1008: Remove aggressive restorer_valid[i] == -1 assert
If we've ever received a signal without an SA_RESTORER flag and with a
non-NULL sa_restorer value, then this assert will trigger when creating
a child thread.  Removing it doesn't appear to break any invariants.

I re-ordered the memset with the installation of our signal info field
to avoid a possible race on fully_initialized, which was for i#26.

Fixes issue 1008.

Review URL: https://codereview.appspot.com/6904055

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1762 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6