i#1061: Check fragment flag compatibility for new unexpected vmareas
When we flush code from non-module vmareas, we remove the area completely.
When we re-execute it, we re-add the area, and we check whether to sandbox
it based on stats in written_areas.  The current flag compatibility check
happens before re-adding the area, but we should also check compatibility
after adding the area.

Fixes issue 1061.

Review URL: https://codereview.appspot.com/7135057

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1844 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6