i#1063: kernel32 Local*()
Implemented the full set of Local*() kernel32 redirection routines.

Added unit tests, which ended up catching bugs in redirect_Rtl*Heap()
routines (to be committed separately) that may explain some
-privlib_privheap bugs we've seen in the past.

Reviewer: it should go without saying, but a reminder to please look at the
man pages and confirm this is a faithful implementation.  Some of the
documentation is a little confusing.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1866 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6