Intercept native cross-module calls with -native_exec_retakeover
This only works for modules with PLT relocations that are resolved lazily, and
therefore call through the resolver inserted in the previous diff.  The
interceptor currently just logs.  Eventually it will begin interpreting from
the target PC.

The special heap stubs cause a leak curiosity when -native_exec_retakeover is
enabled.  This will be fixed in a forthcoming CL which knows how to remove the
PLT hooks.

Review URL: https://codereview.appspot.com/7306048

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1890 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6