i#1063: implement file time routines
Fixed a bug in convert_millis_to_date(): Jan 1, 1601 was a Monday, not a Sunday.

Added convert_date_to_millis(), convert_system_time_to_100ns(), and
convert_100ns_to_system_time().

Implemented redirection for FileTimeToLocalFileTime(),
LocalFileTimeToFileTime(), FileTimeToSystemTime(),
SystemTimeToFileTime(), GetSystemTimeAsFileTime(), GetFileTime(), and
SetFileTime().  GetFileTime() was not on the original import list, but
implementing it made it easier to test the others.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1892 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6