i#1097: Always override DYNAMORIO_OPTIONS on execve
This fixes the behavior change from the new Linux drrun frontend, which
does not set DYNAMORIO_OPTIONS in the environment.  It also prevents the
app from changing the options and then calling execve() on itself to
escape DR, which is a regression since DR was open sourced.

Fixes issue 1097.

Review URL: https://codereview.appspot.com/7388063

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1953 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6