fixes issue 1093
The Windows private loader now computes the GS security cookie and sets it
in all libraries that have the field properly initialized.  The algorithm
for generating the cookie isn't too complex and my implementation is
enough to satisfy bcryptPrimitives.dll, which has a sanity check that it is
appropriately set.

git-svn-id: https://dynamorio.googlecode.com/svn/trunk@1958 49cc7528-f6fd-11dd-9d1a-b59b2e1864b6