i#38: Implement ptrace-based attach for Linux
Fixes an issue with attaching to processes blocked on sleeps or reads.

Adds dr_inject_attach_to_pid() to the drinjectlib API on Linux.

The command line invocation to config and attach is currently:
  ./bin64/drrun -debug -attach <pid> <dr_ops>\
    -c ./suite/tests/bin/libclient.attach.dll.so -- <bogus name>

drrun exits after injecting successfully and does not wait for the app
to terminate.  This suffices for testing, but we should come up with a
better interface for generating a one-time config file and injecting.